+++ src/libsrcml/srcMLUtility.hpp
@@ -118,7 +118,7 @@ private:
     const char* infile;
     const char* output_encoding;
     OPTION_TYPE options;
-    int units;
+    long units;
     const char* diff_version;
     const char * buffer;
     int size;
